chore: update to fix solidJS conflicts, and better scripting for game gallery.
