+++ package.json
@@ -3,12 +3,13 @@
   "type": "module",
   "version": "0.0.1",
   "scripts": {
-    "dev": "npm run generate-mdx && astro dev",
-    "start": "npm run generate-mdx && astro dev",
-    "build": "astro build",
+    "dev": "npm run generate-mdx && npm run generate-game && astro dev",
+    "start": "npm run generate-mdx && npm run generate-game && astro dev",
+    "build": "npm run generate-mdx && npm run generate-game && astro build",
     "preview": "astro preview",
     "astro": "astro",
-    "generate-mdx": "node ./scripts/components.cjs"
+    "generate-mdx": "node ./scripts/components.cjs",
+    "generate-game": "node ./scripts/games.cjs"
   },
   "dependencies": {
     "@astrojs/netlify": "^5.2.0",
@@ -30,7 +31,7 @@
     "remark-frontmatter": "^5.0.0",
     "sharp": "^0.32.6",
     "solid-js": "^1.8.6",
-    "swiper": "^11.1.3",
+    "swiper": "^11.1.4",
     "tailwindcss": "^3.3.5"
   }
 }
