--- src/components/react/GamePage/gamepage.css
@@ -1,61 +1,0 @@
-.game-page-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 2rem;
-  }
-  
-  .page-title {
-    font-size: 2.5rem;
-    margin-bottom: 2rem;
-  }
-  
-  .game-section {
-    width: 100%;
-    margin-bottom: 3rem;
-    background-color: #f5f5f5;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    padding: 2rem;
-  }
-  
-  .game-title {
-    font-size: 2rem;
-    margin-bottom: 1rem;
-  }
-  
-  .game-content {
-    display: flex;
-    align-items: center;
-  }
-  
-  .game-image {
-    width: 300px;
-    height: 200px;
-    object-fit: cover;
-    margin-right: 2rem;
-    border-radius: 8px;
-  }
-  
-  .game-info {
-    flex-grow: 1;
-  }
-  
-  .game-description {
-    font-size: 1.2rem;
-    margin-bottom: 1rem;
-  }
-  
-  .download-button {
-    display: inline-block;
-    background-color: #007bff;
-    color: #fff;
-    text-decoration: none;
-    padding: 0.8rem 1.5rem;
-    border-radius: 4px;
-    transition: background-color 0.3s ease;
-  }
-  
-  .download-button:hover {
-    background-color: #0056b3;
-  }
