+++ src/content/docs/games.mdx
@@ -6,26 +6,28 @@ title: Game Page
 
 
 
+
 ## [Angry Birds](#angry-birds)
 
 ### Description
-A popular mobile game where you launch birds at pigs using a slingshot.
+A simple game of Angry Birds
 
-![Game Image](https://raw.githubusercontent.com/XQuestCode/SplashkitGames/main/angry-birds/template.gif)
+![Game Image for Angry Birds](https://raw.githubusercontent.com/XQuestCode/SplashkitGames/main/angry-birds/template.gif)
 
 
 Language - `cpp` 	
-Rating - `3`
+Rating - `null`
 
 ### Command
 ```cpp
 
-clang++ program.cpp -l SplashKit -o program && ./program
+clang++ *.cpp -l SplashKit -o game && ./game
 ```
 
-<button onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/angry-birds/', '_blank')">Download</button>
+<button class="game-header-button" onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/angry-birds/', '_blank')">Download</button>
 
 ---
+
 
 
 
@@ -38,7 +40,7 @@ A classic arcade game where you pilot a spaceship and destroy asteroids.
 
 
 Language - `csharp` 	
-Rating - `4`
+Rating - `null`
 
 ### Command
 ```csharp
@@ -46,9 +48,10 @@ Rating - `4`
 dotnet run
 ```
 
-<button onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/asteroids/', '_blank')">Download</button>
+<button class="game-header-button" onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/asteroids/', '_blank')">Download</button>
 
 ---
+
 
 
 
@@ -57,11 +60,11 @@ dotnet run
 ### Description
 A classic arcade game where you control a paddle to bounce a ball and break bricks.
 
-![Game Image](https://images.sftcdn.net/images/t_app-cover-l,f_auto/p/0f5173aa-96da-11e6-9405-00163ec9f5fa/4242823519/dx-ball-2-screenshot.jpg)
+![Game Image for DxBallGame](https://raw.githubusercontent.com/XQuestCode/SplashkitGames/main/dx-ball/game.gif)
 
 
 Language - `cpp` 	
-Rating - `4`
+Rating - `null`
 
 ### Command
 ```cpp
@@ -69,30 +72,31 @@ Rating - `4`
 clang++ program.cpp -l SplashKit -o program && ./program
 ```
 
-<button onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/dx-ball/', '_blank')">Download</button>
+<button class="game-header-button" onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/dx-ball/', '_blank')">Download</button>
 
 ---
+
 
 
 
 ## [Road Runner](#road-runner)
 
 ### Description
-A classic arcade game where you control a car and avoid obstacles.
+A basic 2D game where you control a character and collect the stars.
 
 ![](https://images-wixmp-ed30a86b8c4ca887773594c2.wixmp.com/f/760bb8f8-3a05-4e44-be4d-6d1b2a6e5392/d1nno8d-364adddf-a168-44a1-b5e5-db7e2e779331.gif?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1cm46YXBwOjdlMGQxODg5ODIyNjQzNzNhNWYwZDQxNWVhMGQyNmUwIiwiaXNzIjoidXJuOmFwcDo3ZTBkMTg4OTgyMjY0MzczYTVmMGQ0MTVlYTBkMjZlMCIsIm9iaiI6W1t7InBhdGgiOiJcL2ZcLzc2MGJiOGY4LTNhMDUtNGU0NC1iZTRkLTZkMWIyYTZlNTM5MlwvZDFubm84ZC0zNjRhZGRkZi1hMTY4LTQ0YTEtYjVlNS1kYjdlMmU3NzkzMzEuZ2lmIn1dXSwiYXVkIjpbInVybjpzZXJ2aWNlOmZpbGUuZG93bmxvYWQiXX0.SIYcpgo6CrC3ucFcFCIAxRYhR-hFZK_0_NJlE7Z_67I)
 
 
 Language - `cpp` 	
-Rating - `3`
+Rating - `null`
 
 ### Command
 ```cpp
 
-clang++ program.cpp -l SplashKit -o program && ./program
+clang++ *.cpp -l SplashKit -o RoadRunner && ./RoadRunner
 ```
 
-<button onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/road-runner/', '_blank')">Download</button>
+<button class="game-header-button" onclick="window.open('https://download-directory.github.io?url=https://raw.githubusercontent.com/XQuestCode/SplashkitGames/tree/main/road-runner/', '_blank')">Download</button>
 
 ---
