+++ src/content/docs/installation/index.mdx
@@ -9,7 +9,7 @@ import { LinkCard, CardGrid } from "@astrojs/starlight/components";
 
 import Card from "../../../components/Card.astro";
 
-import { ScriptComp } from '../../../assets/ScriptComp'
+import { CardAnimation } from '../../../assets/CardAnimation'
 
 
 <CardGrid>
@@ -51,4 +51,4 @@ import { ScriptComp } from '../../../assets/ScriptComp'
   </Card>
 </CardGrid>
 
-<ScriptComp client:only />
+<CardAnimation client:only />
